chore: 빌드 에러 : 경로 추가
