+++ src/features/find/components/filters/FindPrivateFilter.tsx
@@ -2,76 +2,80 @@
 import { useState } from 'react'
 import Image from 'next/image'
 import { useRouter, useSearchParams } from 'next/navigation'
-import TeamFilterModal from './modal/TeamFilterModal'
+import PrivateFilterModal from '../modal/PrivateFilterModal'
 
-export default function FindTeamFilter() {
+export default function FindPrivateFilter() {
   const router = useRouter()
   const searchParams = useSearchParams()
 
   const [isFilterOpen, setIsFilterOpen] = useState(false)
 
   // URL에서 필터 상태 가져오기 - 여러 값을 배열로 가져오도록 수정
-  const selectedSize = searchParams.getAll('scaleName')
+  const selectedPositions = searchParams.getAll('subPosition')
   const selectedLocations = searchParams.getAll('cityName')
-  const selectedStatus = searchParams.getAll('teamStateName')
+  const selectedStatus = searchParams.getAll('profileStateName')
 
   // 필터 적용 핸들러
-  const handleApplyFilters = (filters: { scaleNames: string[]; cityNames: string[]; teamStateNames: string[] }) => {
+  const handleApplyFilters = (filters: {
+    subPositions: string[]
+    cityNames: string[]
+    profileStateNames: string[]
+  }) => {
     updateURLParams(filters)
   }
 
   // URL 파라미터 업데이트 함수
-  const updateURLParams = (filters: { scaleNames: string[]; cityNames: string[]; teamStateNames: string[] }) => {
+  const updateURLParams = (filters: { subPositions: string[]; cityNames: string[]; profileStateNames: string[] }) => {
     const params = new URLSearchParams()
 
     // 각 필터 타입별로 여러 값을 추가
-    filters.scaleNames.forEach((size) => {
-      params.append('scaleName', size)
+    filters.subPositions.forEach((position) => {
+      params.append('subPosition', position)
     })
 
     filters.cityNames.forEach((city) => {
       params.append('cityName', city)
     })
 
-    filters.teamStateNames.forEach((state) => {
-      params.append('teamStateName', state)
+    filters.profileStateNames.forEach((state) => {
+      params.append('profileStateName', state)
     })
 
-    params.set('page', '0')
-    router.push(`/find/team?${params.toString()}`)
+    params.set('page', '1')
+    router.push(`/find/private?${params.toString()}`)
   }
 
   // 개별 필터 제거 핸들러
-  const removeSize = (size: string) => {
+  const removePosition = (position: string) => {
     updateURLParams({
-      scaleNames: selectedSize.filter((s) => s !== size),
+      subPositions: selectedPositions.filter((p) => p !== position),
       cityNames: selectedLocations,
-      teamStateNames: selectedStatus,
+      profileStateNames: selectedStatus,
     })
   }
 
   const removeLocation = (location: string) => {
     updateURLParams({
-      scaleNames: selectedSize,
+      subPositions: selectedPositions,
       cityNames: selectedLocations.filter((l) => l !== location),
-      teamStateNames: selectedStatus,
+      profileStateNames: selectedStatus,
     })
   }
 
   const removeStatus = (status: string) => {
     updateURLParams({
-      scaleNames: selectedSize,
+      subPositions: selectedPositions,
       cityNames: selectedLocations,
-      teamStateNames: selectedStatus.filter((s) => s !== status),
+      profileStateNames: selectedStatus.filter((s) => s !== status),
     })
   }
 
   // 필터 초기화 핸들러
   const resetFilters = () => {
     updateURLParams({
-      scaleNames: [],
+      subPositions: [],
       cityNames: [],
-      teamStateNames: [],
+      profileStateNames: [],
     })
   }
 
@@ -92,8 +96,8 @@ export default function FindTeamFilter() {
               onClick={() => setIsFilterOpen(true)}
               className="flex cursor-pointer flex-col  gap-2 rounded-xl border border-grey30 px-5 py-4 text-sm hover:bg-[#EDF3FF]"
             >
-              <p className="flex justify-center text-grey70 md:justify-start">규모</p>
-              <p className="hidden text-grey50 md:flex">선호하는 팀 규모를 선택해 주세요</p>
+              <p className="flex justify-center text-grey70 md:justify-start">포지션</p>
+              <p className="hidden text-grey50 md:flex">포지션을 선택해 주세요</p>
             </div>
             <div
               onClick={() => setIsFilterOpen(true)}
@@ -107,20 +111,20 @@ export default function FindTeamFilter() {
               className="flex cursor-pointer flex-col gap-2 rounded-xl border border-grey30 px-5 py-4 text-sm hover:bg-[#EDF3FF]"
             >
               <p className="flex justify-center text-grey70 md:justify-start">현재 상태</p>
-              <p className="hidden text-grey50 md:flex">어떤 팀을 찾고 있는지 선택해 주세요</p>
+              <p className="hidden text-grey50 md:flex">어떤 팀원을 찾고 있는지 선택해 주세요</p>
             </div>
           </div>
 
           {/* 선택된 필터들 표시 */}
-          {(selectedSize.length > 0 || selectedLocations.length > 0 || selectedStatus.length > 0) && (
+          {(selectedPositions.length > 0 || selectedLocations.length > 0 || selectedStatus.length > 0) && (
             <div className="mt-2 flex w-full items-center gap-2 overflow-x-auto">
-              {selectedSize.map((size) => (
+              {selectedPositions.map((position) => (
                 <div
-                  key={size}
-                  onClick={() => removeSize(size)}
+                  key={position}
+                  onClick={() => removePosition(position)}
                   className="flex shrink-0 cursor-pointer items-center gap-2 rounded-[0.25rem] bg-grey10 px-3 py-2"
                 >
-                  <span className="text-sm text-main">{size}</span>
+                  <span className="text-sm text-main">{position}</span>
                   <Image src="/common/icons/delete_icon.svg" alt="close" width={16} height={16} />
                 </div>
               ))}
@@ -150,14 +154,14 @@ export default function FindTeamFilter() {
       </div>
 
       {isFilterOpen && (
-        <TeamFilterModal
+        <PrivateFilterModal
           isFilterOpen={isFilterOpen}
           setIsFilterOpen={setIsFilterOpen}
           onApplyFilters={handleApplyFilters}
           initialFilters={{
-            scaleNames: selectedSize,
+            subPositions: selectedPositions,
             cityNames: selectedLocations,
-            teamStateNames: selectedStatus,
+            profileStateNames: selectedStatus,
           }}
         />
       )}
