+++ src/features/find/components/filters/FindTeamFilter.tsx
@@ -2,80 +2,76 @@
 import { useState } from 'react'
 import Image from 'next/image'
 import { useRouter, useSearchParams } from 'next/navigation'
-import PrivateFilterModal from './modal/PrivateFilterModal'
+import TeamFilterModal from '../modal/TeamFilterModal'
 
-export default function FindPrivateFilter() {
+export default function FindTeamFilter() {
   const router = useRouter()
   const searchParams = useSearchParams()
 
   const [isFilterOpen, setIsFilterOpen] = useState(false)
 
   // URL에서 필터 상태 가져오기 - 여러 값을 배열로 가져오도록 수정
-  const selectedPositions = searchParams.getAll('subPosition')
+  const selectedSize = searchParams.getAll('scaleName')
   const selectedLocations = searchParams.getAll('cityName')
-  const selectedStatus = searchParams.getAll('profileStateName')
+  const selectedStatus = searchParams.getAll('teamStateName')
 
   // 필터 적용 핸들러
-  const handleApplyFilters = (filters: {
-    subPositions: string[]
-    cityNames: string[]
-    profileStateNames: string[]
-  }) => {
+  const handleApplyFilters = (filters: { scaleNames: string[]; cityNames: string[]; teamStateNames: string[] }) => {
     updateURLParams(filters)
   }
 
   // URL 파라미터 업데이트 함수
-  const updateURLParams = (filters: { subPositions: string[]; cityNames: string[]; profileStateNames: string[] }) => {
+  const updateURLParams = (filters: { scaleNames: string[]; cityNames: string[]; teamStateNames: string[] }) => {
     const params = new URLSearchParams()
 
     // 각 필터 타입별로 여러 값을 추가
-    filters.subPositions.forEach((position) => {
-      params.append('subPosition', position)
+    filters.scaleNames.forEach((size) => {
+      params.append('scaleName', size)
     })
 
     filters.cityNames.forEach((city) => {
       params.append('cityName', city)
     })
 
-    filters.profileStateNames.forEach((state) => {
-      params.append('profileStateName', state)
+    filters.teamStateNames.forEach((state) => {
+      params.append('teamStateName', state)
     })
 
-    params.set('page', '1')
-    router.push(`/find/private?${params.toString()}`)
+    params.set('page', '0')
+    router.push(`/find/team?${params.toString()}`)
   }
 
   // 개별 필터 제거 핸들러
-  const removePosition = (position: string) => {
+  const removeSize = (size: string) => {
     updateURLParams({
-      subPositions: selectedPositions.filter((p) => p !== position),
+      scaleNames: selectedSize.filter((s) => s !== size),
       cityNames: selectedLocations,
-      profileStateNames: selectedStatus,
+      teamStateNames: selectedStatus,
     })
   }
 
   const removeLocation = (location: string) => {
     updateURLParams({
-      subPositions: selectedPositions,
+      scaleNames: selectedSize,
       cityNames: selectedLocations.filter((l) => l !== location),
-      profileStateNames: selectedStatus,
+      teamStateNames: selectedStatus,
     })
   }
 
   const removeStatus = (status: string) => {
     updateURLParams({
-      subPositions: selectedPositions,
+      scaleNames: selectedSize,
       cityNames: selectedLocations,
-      profileStateNames: selectedStatus.filter((s) => s !== status),
+      teamStateNames: selectedStatus.filter((s) => s !== status),
     })
   }
 
   // 필터 초기화 핸들러
   const resetFilters = () => {
     updateURLParams({
-      subPositions: [],
+      scaleNames: [],
       cityNames: [],
-      profileStateNames: [],
+      teamStateNames: [],
     })
   }
 
@@ -96,8 +92,8 @@ export default function FindPrivateFilter() {
               onClick={() => setIsFilterOpen(true)}
               className="flex cursor-pointer flex-col  gap-2 rounded-xl border border-grey30 px-5 py-4 text-sm hover:bg-[#EDF3FF]"
             >
-              <p className="flex justify-center text-grey70 md:justify-start">포지션</p>
-              <p className="hidden text-grey50 md:flex">포지션을 선택해 주세요</p>
+              <p className="flex justify-center text-grey70 md:justify-start">규모</p>
+              <p className="hidden text-grey50 md:flex">선호하는 팀 규모를 선택해 주세요</p>
             </div>
             <div
               onClick={() => setIsFilterOpen(true)}
@@ -111,20 +107,20 @@ export default function FindPrivateFilter() {
               className="flex cursor-pointer flex-col gap-2 rounded-xl border border-grey30 px-5 py-4 text-sm hover:bg-[#EDF3FF]"
             >
               <p className="flex justify-center text-grey70 md:justify-start">현재 상태</p>
-              <p className="hidden text-grey50 md:flex">어떤 팀원을 찾고 있는지 선택해 주세요</p>
+              <p className="hidden text-grey50 md:flex">어떤 팀을 찾고 있는지 선택해 주세요</p>
             </div>
           </div>
 
           {/* 선택된 필터들 표시 */}
-          {(selectedPositions.length > 0 || selectedLocations.length > 0 || selectedStatus.length > 0) && (
+          {(selectedSize.length > 0 || selectedLocations.length > 0 || selectedStatus.length > 0) && (
             <div className="mt-2 flex w-full items-center gap-2 overflow-x-auto">
-              {selectedPositions.map((position) => (
+              {selectedSize.map((size) => (
                 <div
-                  key={position}
-                  onClick={() => removePosition(position)}
+                  key={size}
+                  onClick={() => removeSize(size)}
                   className="flex shrink-0 cursor-pointer items-center gap-2 rounded-[0.25rem] bg-grey10 px-3 py-2"
                 >
-                  <span className="text-sm text-main">{position}</span>
+                  <span className="text-sm text-main">{size}</span>
                   <Image src="/common/icons/delete_icon.svg" alt="close" width={16} height={16} />
                 </div>
               ))}
@@ -154,14 +150,14 @@ export default function FindPrivateFilter() {
       </div>
 
       {isFilterOpen && (
-        <PrivateFilterModal
+        <TeamFilterModal
           isFilterOpen={isFilterOpen}
           setIsFilterOpen={setIsFilterOpen}
           onApplyFilters={handleApplyFilters}
           initialFilters={{
-            subPositions: selectedPositions,
+            scaleNames: selectedSize,
             cityNames: selectedLocations,
-            profileStateNames: selectedStatus,
+            teamStateNames: selectedStatus,
           }}
         />
       )}
